# src/modules/provider/dto/create-provider.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsOptional } from 'class-validator';

export class CreateProviderDto {
    @ApiProperty()
    @IsNotEmpty({ message: 'Tên nhà cung cấp không được để trống' })
    name: string;

    @ApiProperty()
    @IsOptional()
    address: string;

    @ApiProperty()
    @IsOptional()
    phone: string;

    @ApiProperty()
    @IsOptional()
    email: string;

    @ApiProperty()
    @IsOptional()
    description: string;
}