# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/permission": {
      "post": {
        "operationId": "PermissionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Permission"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "get": {
        "operationId": "PermissionController_findAll",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "perPage",
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "id.ASC",
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Permission"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/permission/{id}": {
      "get": {
        "operationId": "PermissionController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Permission"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "operationId": "PermissionController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePermissionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Permission"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "PermissionController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Permission"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/media/upload": {
      "post": {
        "operationId": "MediaController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/media": {
      "get": {
        "operationId": "MediaController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/media/{id}": {
      "get": {
        "operationId": "MediaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "MediaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Media"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
    "/department": {
      "post": {
        "operationId": "DepartmentController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Department"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "get": {
        "operationId": "DepartmentController_findAll",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "perPage",
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "id.ASC",
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Department"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/department/{id}": {
      "get": {
        "operationId": "DepartmentController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Department"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "operationId": "DepartmentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Department"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "DepartmentController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Department"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
...
    "/provider": {
      "post": {
        "operationId": "ProviderController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "get": {
        "operationId": "ProviderController_findAll",
        "parameters": [
          {
            "required": false,
            "name": "page",
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "perPage",
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "required": false,
            "name": "sortBy",
            "in": "query",
            "schema": {
              "default": "id.ASC",
              "type": "string"
            }
          },
          {
            "required": false,
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    },
    "/provider/{id}": {
      "get": {
        "operationId": "ProviderController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "patch": {
        "operationId": "ProviderController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      },
      "delete": {
        "operationId": "ProviderController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Provider"
        ],
        "security": [
          {
            "authorization": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Warehouse MS Swagger",
    "description": "The Warehouse MS API documents",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "authorization": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    },
    "schemas": {
...
      "CreatePermissionDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePermissionDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateDepartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description"
        ]
      },
      "UpdateDepartmentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
      "CreateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "address",
          "phone",
          "email",
          "description"
        ]
      },
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      }
    }
  }
}